+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 
 * Added command `list-service-tags`.
 * `dns zone import`: Fix issue where users could not import wildcard A records.
+* `watcher flow-log configure`: Fixed issue where flow logging could not be enabled in certain regions.
 
 2.5.0
 +++++