+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -7,11 +7,14 @@ Release History
 * `dns zone import`: Fix issue where users could not import wildcard A records.
 * `watcher flow-log configure`: Fixed issue where flow logging could not be enabled in certain regions.
 
+2.5.1
++++++
+* `application-gateway`: Added `waf-policy` commands to support custom WAF rules.
+* `application-gateway create/update`: Added `--waf-policy` and `--max-capacity` arguments.
+
 2.5.0
 +++++
 * [BREAKING CHANGE]: previously deprecated command group `network interface-endpoints` has been removed. Use `network private-endpoints` instead.
-* `application-gateway`: Added `waf-policy` commands to support custom WAF rules.
-* `application-gateway create/update`: Added `--waf-policy` and `--max-capacity` arguments.
 * `network vnet subnet create/update`: Added `--nat-gateway` argument for attaching to a NAT gateway.
 * `dns zone import`: Fix issue where record names could not match a record type.